T.5: daemon hardening — runtime status-cache cap, reconcile bounds, peer retry, shutdown (#240)
* test: add windows runtime parity coverage

* feat: enforce immutable sqlite message rows

* fix: close T.4 qa findings

* docs: note T.5 accepted accept-error injection

* Fix T.4 Windows shutdown test flake

* Fix T.3 immutable row QA findings

* Fix T.3 QA-3 batch isolation tests

* Fix T.4 QA-4 immutable rows sync

* Fix T.3 QA-4 duplicate deadline flush test

* fix: close T.4 QA-5 rusqlite gaps

* fix: close T.3 QA-5 writer cleanup

* fix: remove sqlite connection helper boundary leak

* feat: finish t5 daemon hardening

* fix: close T.5 QA-1 hardening findings

* fix: guard against None stdout/stderr in run_lint.py on Windows Python 3.14

Same fix as T.7: add errors="replace" to subprocess.run and null-guard all
stdout/stderr references to prevent AttributeError on Windows Python 3.14
when a lint tool emits non-UTF-8 bytes.


* fix: name lifecycle join-helper thread (RSH-T5-NEW-002)

thread::Builder::new().name("atm-daemon-lifecycle-join-helper") in
shutdown_worker_with_timeout at lifecycle_control.rs:253, consistent
with the named-thread pattern applied to composition.rs in the prior fix.


---------
