+++ crates/atm-daemon/src/lib.rs
@@ -15,6 +15,7 @@ mod host_ownership;
 mod lifecycle_control;
 mod local_ipc_connection;
 mod local_ipc_transport;
+mod local_ipc_wake;
 mod notification_runtime;
 mod peer_transport;
 mod reconcile_runtime;
